+++ tests/ui/invalid_sel.rs
@@ -1,0 +1,8 @@
+//! Test invalid selector syntax
+use objc2::sel;
+
+fn main() {
+    sel!();
+    sel!(a: b);
+    sel!(a: b: c);
+}